+++ gems/aws-sdk-core/lib/aws-sdk-core/instance_profile_credentials.rb
@@ -90,6 +90,7 @@ module Aws
       @token = nil
       @no_refresh_until = nil
       @async_refresh = false
+      @metrics = ['CREDENTIALS_IMDS']
       super
     end